requirements.txt
# Core dependencies
pyyaml>=6.0
numpy>=1.22.0
matplotlib>=3.5.0

# Testing
pytest>=7.0.0
pytest-cov>=4.1.0

# Development
black>=23.1.0
isort>=5.12.0
flake8>=6.0.0

# Documentation
sphinx>=6.1.3
sphinx-rtd-theme>=1.2.0
